Update menu.js
from slack: So i've made it so the map data can be saved onto localStorage, just trying to figure out how to push to github....just give me 4 or 5 hour for that. A few notes about my addition, some of it is waiting on other stuff to be done. Right now I have it working with data I entered manually with variables I created for the temporary map I created. Once we get the map building off of map_file and updating the hero info and map info, we can, obviously, change my code to reflect what is used by the map builder. Right now, all the information is stored in an array in JSON format so when you import the map data you should (ideally) be able to use the JSON.parse(localStorage(saved_map_info)) and should store everything... I think. If you have any questions feel free to ask as I will probably start playing around with other stuff